# Use sets to improve performance of dfs search