@xadupre 2024-07-04
It seems reachable + unreachable = set of nodes. Do you it is doable to remoave unreachable and simplify the function signature assuming performance stays the same?