@kylesayrs 2024-07-05
@xadupre The `reachable` set must be maintained, as it is the ultimate purpose of the function. This leaves the `unreachable` set to be removed. Unfortunately, in order to iterate through the negative set of `reachable`, the function must iterate through all node indices.

```python3
nodes_to_search = [
    index
    for index in range(len(nodes))
    if node_output_name in nodes[index].output and index not in reachable
]
```

This leads to a longer runtime for larger models

| Model Name | Num Nodes | Previous | New | One Set |
| ---- | -------- | --------- | ----- | --- |
| obertquant.onnx | 1271  | 0.158s   | 0.110s | 0.118s |
| ai-town-3B.onnx | 3515  | 8.002s  | 3.725s | 6.094s |